move error out of trait
