+++ crates/runtime/Cargo.toml
@@ -12,7 +12,7 @@ description = "The spiced runtime"
 snafu.workspace = true
 spicepod = { path = "../spicepod" }
 app = { path = "../app" }
-axum = { version = "0.7.4"}
+axum = { version = "0.7.4", features = ["macros"]}
 tokio.workspace = true
 tracing.workspace = true
 clap.workspace = true
